+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/ai_guard/requests/TextGuardRequest.java
@@ -17,11 +17,32 @@ import lombok.extern.jackson.Jacksonized;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 @SuperBuilder
 @Value
-public final class TextGuardRequest extends BaseRequest {
+public final class TextGuardRequest<TMessages> extends BaseRequest {
 
-	@NonNull
+	/**
+	 * Text to be scanned by AI Guard for PII, sensitive data, malicious
+	 * content, and other data types defined by the configuration. Supports
+	 * processing up to 10KB of text.
+	 */
 	String text;
 
+	/**
+	 * Structured data to be scanned by AI Guard for PII, sensitive data,
+	 * malicious content, and other data types defined by the configuration.
+	 * Supports processing up to 10KB of text.
+	 */
+	TMessages messages;
+
+	/**
+	 * Recipe key of a configuration of data types and settings defined in the
+	 * Pangea User Console. It specifies the rules that are to be applied to the
+	 * text, such as defang malicious URLs.
+	 */
 	String recipe;
+
+	/**
+	 * Setting this value to true will provide a detailed analysis of the text
+	 * data
+	 */
 	boolean debug;
 }