+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/ai_guard/responses/TextGuardResponse.java
@@ -3,4 +3,4 @@ package cloud.pangeacyber.pangea.ai_guard.responses;
 import cloud.pangeacyber.pangea.Response;
 import cloud.pangeacyber.pangea.ai_guard.results.TextGuardResult;
 
-public final class TextGuardResponse extends Response<TextGuardResult> {}
+public final class TextGuardResponse<TMessages> extends Response<TextGuardResult<TMessages>> {}