+++ CHANGELOG.md
@@ -10,11 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Prompt Guard: `classify` and `threshold` parameters.
+- AI Guard: support for structured input.
+- AI Guard: more detectors.
 
 ### Changed
 
 - Deprecated `Config.enviroment` (typo) in favor of `Config.environment`.
 - Clarified what `Config.environment` affects.
+- AI Guard: replaced `redacted` with `action`.
 
 ## 4.4.0-beta-2 - 2025-01-21